include retune messages when exporting/saving midi file
options to export pb-messages added to relevant
functions (save-as-midi, export to midifile, save-midifile).  Existing
behavior (no pb-messages exported) is default for now.

TODO: clarify regime for midi-retuning (perhaps it is clear?),
possibly add explicit choice for exporting pb-messages in prefs, have
this fit with the chosen auto-retune regime (if possible)